材質非表示設定復活

+++ pkg/usecase/preview_usecase.go
@@ -44,7 +44,7 @@ func createPreviewVmd(outputModel *pmx.PmxModel, blurRootBone, blurBone *pmx.Bon
 			angle := float64(j) * 90.0 * flag
 
 			index := (i * 6 * 20) + (j * 20)
-			bf := vmd.NewBoneFrame(index)
+			bf := vmd.NewBoneFrame(float32(index))
 			// 少し上に表示する
 			bf.Position = &mmath.MVec3{X: 0.0, Y: 3.0, Z: 0.0}
 			// 回す方向はブレの軸制限方向
